SalaryService is OK.

+++ hr/src/main/java/hu/webuni/hr/kinela/service/EmployeeService.java
@@ -3,5 +3,5 @@ package hu.webuni.hr.kinela.service;
 import hu.webuni.hr.kinela.model.Employee;
 
 public interface EmployeeService {
-	public int getPayRaisePercent(Employee employee);
+	public float getPayRaisePercent(Employee employee);
 }
